+++ app/enterprise.module.js
@@ -1,5 +1,0 @@
-'use strict';
-
-// for loading styles we need to load main scss file
-import './styles/styles-enterprise.scss';
-