+++ webpack.prod.config.js
@@ -14,7 +14,6 @@ module.exports = {
   entry: {
     'vendor': './app/vendor.module.js',
     'app': './app/app.module.js',
-    'enterprise': './app/enterprise.module.js'
   },
   output: {
     filename: 'static/libs/[name].bundle.js',
@@ -29,7 +28,6 @@ module.exports = {
       template: './app/index.tpl.html',
       inject: 'head',
       filename: 'index.html',
-      excludeAssets: [/enterprise.css/]
     }),
     new CopyWebpackPlugin([
       {