+++ CHANGES.rst
@@ -6,7 +6,10 @@ Changes for Crate Admin Interface
 Unreleased
 ==========
 
+- Align with license change. Enterprise features are now available to everyone.
+
 - Fixed translations. French now works again.
+
 - Fixed missing whitespace.